-(agent/mibgroup/mibII/interfaces.c snmplib/snmpusm.c) fixes to build HP & AIX
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2618 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/mibII/interfaces.c
@@ -942,9 +942,9 @@ var_ifEntry(struct variable *vp,
 u_char *
 var_ifEntry(struct variable *vp,
 	    oid *name,
-	    int *length,
+	    size_t *length,
 	    int exact,
-	    int *var_len,
+	    size_t *var_len,
 	    WriteMethod **write_method)
 {
     static struct ifnet ifnet;
